Remove change in main.m and change submodule
- Remove changes in main.m, turns out we cannot actually catch it here.
Tried some other methods, however none worked.
- Change dupeguru submodule to a more "known" commit that did not have
partial work in it.

NOTE: Issue with picture mode appears to be from python 3.8+ changing to use
`spawn` as the method for multiprocessing.  While this is most likely better
we currently have issues when application is run normally (running from
run.py does not have issues). I am manually going to change to `fork`
instead since other methods did not work. This needs to be done on line
20 of matchblock.py.  Either need to
find a workaround or make a way to detect we are running with cocoa vs.
qt to make this switch.  Just on macos
would affect qt build which does not
need this change.